ghost-cyber-universe-research/requirements.txt
# Ghost Research Core - Unified Requirements

# Framework Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
chromadb==0.4.18
faiss-cpu==1.7.4

# IA & NLP
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.0.0
spacy==3.7.2

# Security & Protection LLM
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
cryptography==41.0.8
argon2-cffi==23.1.0

# Scraping & RSS
feedparser==6.0.10
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Traduction
deep-translator==1.11.4
langdetect==1.0.9

# Analyse de sécurité Cyber
cvss==2.6
python-dateutil==2.8.2

# Elasticsearch
elasticsearch==8.11.1
elasticsearch-async==6.2.0

# Notifications
firebase-admin==6.3.0
python-telegram-bot==20.7

# Utilitaires & Logging
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
schedule==1.2.0
pytz==2023.3
python-slugify==8.0.1
loguru==0.7.2
structlog==23.2.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==7.0.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==24.8.0
ruff==0.6.9
bandit==1.7.9
